Extract postLink function
